fix: 로또 당첨 개수 반환 로직을 WinningLotto로 이동
