+++ src/main/kotlin/lotto/service/WinningStatics.kt
@@ -1,20 +1,10 @@
 package lotto.service
 
 import lotto.model.LottoPrize
-import lotto.model.LottoStore
 import lotto.model.PurchaseInfo
-import lotto.model.WinningLotto
 import kotlin.math.floor
 
-object ResultCalculator {
-    fun calculatePrizeCount(
-        lottoStore: LottoStore,
-        winningLotto: WinningLotto,
-    ) = lottoStore.lottos
-        .map { lotto -> lotto.compare(winningLotto.lotto, winningLotto.bonusNumber) }
-        .groupBy { it }
-        .mapValues { it.value.size }
-
+class WinningStatics {
     fun calculateProfitRatio(
         purchaseInfo: PurchaseInfo,
         prizeCount: Map<LottoPrize, Int>,
