+++ src/test/kotlin/lotto/service/WinningStaticsTest.kt
@@ -1,0 +1,38 @@
+package lotto.service
+
+import lotto.model.LottoPrize
+import lotto.model.PurchaseInfo
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.Arguments
+import org.junit.jupiter.params.provider.MethodSource
+
+class WinningStaticsTest {
+    @ParameterizedTest
+    @MethodSource("로또 수익률 계산 테스트 데이터")
+    fun `로또 수익률을 계산한다`(
+        lottoPrice: String,
+        prizeCount: Map<LottoPrize, Int>,
+        profitRatio: Double,
+    ) {
+        // given
+        val purchaseInfo = PurchaseInfo(lottoPrice)
+        val winningStatics = WinningStatics()
+
+        // when
+        val actual = winningStatics.calculateProfitRatio(purchaseInfo, prizeCount)
+
+        // then
+        assertThat(actual).isEqualTo(profitRatio)
+    }
+
+    companion object {
+        @JvmStatic
+        fun `로또 수익률 계산 테스트 데이터`() =
+            listOf(
+                Arguments.of("14000", mapOf(LottoPrize.FIFTH to 1), 0.35),
+                Arguments.of("8000", mapOf(LottoPrize.FIFTH to 1), 0.62),
+                Arguments.of("1000", mapOf(LottoPrize.FIRST to 1), 2_000_000),
+            )
+    }
+}
